satisfies keyword and file separate
